chore(test-util): use generic steps where possible
