+++ test-util/src/main/java/io/zeebe/test/util/bpmn/random/steps/StepActivateBPMNElement.java
@@ -1,3 +1,10 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
+ */
 package io.zeebe.test.util.bpmn.random.steps;
 
 import java.time.Duration;
